qmi: add qmi_{enc,de}code_message2
This is an updated API for use with upcoming qmic improvements, it
encodes the message type and ID, as well as a pointer to the associated
struct qmi_elem_info array directly from the QMI IDL by generating
macros to initialise the message struct.

This should make message handling a lot cleaner for users.